Create SkeletonWrapper and topicItem skeleton
